% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgExif.R
\name{imgExif}
\alias{imgExif}
\title{Read Exif from one mission}
\usage{
imgExif(imgPath)
}
\arguments{
\item{imgPath}{directory with the images}
}
\value{
sf points of image positions
}
\description{
Reads the exif data from all images of a flight and returns them as points
}
\author{
Marvin Ludwig
}
