% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgGrid.R
\name{imgGrid}
\alias{imgGrid}
\title{Sample Image Grid}
\usage{
imgGrid(imgBB, gridsize)
}
\arguments{
\item{imgBB}{the bounding box to sample}

\item{gridsize}{distance between the images}
}
\value{
points with regular image location
}
\description{
Samples a regular grid with equal distance between the images
}
\author{
Marvin Ludwig
}
