% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initL0.R
\name{initL0}
\alias{initL0}
\title{Creates Level 0 structure}
\usage{
initL0(pathL0)
}
\arguments{
\item{pathL0}{[string] A directory path}
}
\description{
creates directories for L0 image data
}
\author{
Marvin Ludwig
}
