% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointArea.R
\name{minBB}
\alias{minBB}
\title{Rectangle flight area around points}
\usage{
minBB(points, buffer = 20, epsg = 25832)
}
\arguments{
\item{points}{a sp object, points you wanr to fly over}
}
\value{
SpatialPoints: Corners of the flight area
}
\description{
Creates rectangle flight area based on SpatialPoints
}
\details{
The code is based on a Rotating Caliper Algorithm and mostly copy and pasted (see reference)
}
\references{
http://dwoll.de/rexrepos/posts/diagBounding.html
}
\author{
Marvin Ludwig
}
