% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRAW.R
\docType{data}
\name{path}
\alias{path}
\title{Read Exif from one mission}
\format{An object of class \code{character} of length 1.}
\usage{
path
}
\arguments{
\item{path}{}
}
\value{

}
\description{

}
\author{
Marvin Ludwig
}
\keyword{datasets}
