% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonFP.R
\name{polygonFlight}
\alias{polygonFlight}
\title{Polygon to FP line}
\usage{
polygonFlight(p, sides = c(4, 3), take_off = c(477704, 5632173),
  epsg = 25832)
}
\description{
converts a polygon to a line used by makeAP
}
\author{
Marvin Ludwig
}
