% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2L0.R
\name{raw2L0}
\alias{raw2L0}
\title{Analyses raw image UAV data and returns a L0 report}
\usage{
raw2L0(pathRAW, taxi_threshold = c(3, 5))
}
\arguments{
\item{pathRAW}{[string] path to the raw images}

\item{taxi_threshold}{[numeric] vector of length 2 with the lower and upper distance between two consecutive images}
}
\value{
list of exif
}
\description{
Reads the EXIF information and divides the images into noGPS, taxi and useable L0 images
}
\details{
Creates a subfolder in pathRAW with the following structure:
├── flightArea
├── flightLog
├── imageReport
│   ├── images_l0.csv
│   ├── images_noGPS.csv
│   └── images_taxi.csv
├── images
└── trashcan
├── noGPS
└── taxi
Currently doesn't move any files
}
\author{
Marvin Ludwig
}
