% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTasks.R
\name{readTask}
\alias{readTask}
\title{Read UAV Task}
\usage{
readTask(taskfile)
}
\arguments{
\item{taskfile}{filepath of the task}
}
\description{
Reads a UAV mavlink flight task and converts to sf points
}
\author{
Marvin Ludwig
}
