% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleImages.R
\name{sampleImages}
\alias{sampleImages}
\title{Sample Images}
\usage{
sampleImages(images, task, spacing = 20)
}
\arguments{
\item{images}{bla}
}
\value{

}
\description{
Images at sample poitns
}
\author{
Marvin Ludwig
}
