% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleTask.R
\name{sampleTask}
\alias{sampleTask}
\title{Sample points along the flight path}
\usage{
sampleTask(task, spacing = 20)
}
\arguments{
\item{task}{sf points from \code{\link[readTask]{readTask}}}

\item{spacing}{distance between the sample points}
}
\value{

}
\description{
Samples regular spaced points along the UAV flight path
}
\author{
Marvin Ludwig
}
