% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskRead.R
\name{taskRead}
\alias{taskRead}
\title{Read UAV Task}
\usage{
taskRead(taskfile)
}
\arguments{
\item{taskfile}{filepath of the task}
}
\description{
Reads a UAV mavlink flight task and converts to sf points
}
\author{
Marvin Ludwig
}
